\name{G13Lab03-package}
\alias{G13Lab03-package}
\alias{G13Lab03}
\docType{package}
\title{
\packageTitle{G13Lab03}
}
\description{
\packageDescription{G13Lab03}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{G13Lab03}
\packageIndices{G13Lab03}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{G13Lab03}

Maintainer: \packageMaintainer{G13Lab03}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
