% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Six Nodes and their cost to achieve destination from source node.}
\format{
A data frame with 19 rows and 3 variables:
\describe{
\item{v1}{Edge of the graph, which considered as source node}
\item{v2}{Edge of the graph, which considered as destination node}
\item{w}{weight of the edge from V1 to V2}
}
}
\usage{
wiki_graph
}
\description{
A data.frame with three variables (v1, v2 and w) that contains
the edges of the graph (from v1 to v2) with the weight of the edge (w)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Graph_(mathematics}
}
\keyword{datasets}
