\name{G13Lab03-package}
\alias{G13Lab03-package}
\alias{G13Lab03}
\docType{package}
\title{
\packageTitle{G13Lab03}
}
\description{
\packageDescription{G13Lab03}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{G13Lab03}
\packageIndices{G13Lab03}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{G13Lab03}

Maintainer: \packageMaintainer{G13Lab03}
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
https://en.wikipedia.org/wiki/Graph_(mathematics, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
